+++ lib/jose.dart
@@ -7,3 +7,4 @@ export 'src/jwk.dart';
 export 'src/jws.dart';
 export 'src/jwe.dart';
 export 'src/jwt.dart';
+export 'src/jwa.dart';